完善初始化脚本和修复 bug

+++ env_init/src/uninstall_vault.sh
@@ -12,12 +12,8 @@ if [[ $1 = -n ]]; then noenter=true; else noenter=false; fi
 echo -e "${b}------------------- Vault 卸载 -------------------${n}"
 
 echo -e "${b}正在卸载 Vault...${n}"
-sudo systemctl stop vault
-expect << EOF
-spawn sudo apt-get remove --purge vault
-expect "Do you want to continue?" { send "Y\r" }
-expect eof
-EOF
+if sudo systemctl is-active --quiet vault; then sudo systemctl stop vault; fi
+sudo apt-get remove --purge vault -y
 sudo rm -rf /etc/vault.d
 if [[ $noenter = true ]]; then ans=y
 else echo -e -n "${r}删除 Vault 的数据？(y/n): ${n}"; read -p "" ans; fi
@@ -25,10 +21,10 @@ if [[ $ans = y ]]; then
     echo -e "${b}删除所有 Vault 的数据.${n}"
     sudo rm -rf /opt/vault
 fi
+
 sudo systemctl daemon-reload
 
-rm -f /root/.env
-
-echo -e "${g}Vault 已成功卸载。${n}"
+echo -e "${b}删除 medusa...${n}"
+rm -rf /home/woo/.medusa
 
 echo -e "${b}------------------- Vault 卸载 完成 -------------------${n}"
